% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_wgts.R
\name{append_attr_wgts_to_imp_data}
\alias{append_attr_wgts_to_imp_data}
\title{Add Attrition Weights to Imputed Data}
\usage{
append_attr_wgts_to_imp_data(data.dir, attr.dir)
}
\arguments{
\item{data.dir}{a character defining where recoded imputed data is}

\item{attr.dir}{a character defining where the attrition weight results are}
}
\value{
null but resave recoded imputed data into data.dir
}
\description{
TO-DO
}
\details{
Adds estimated attrition weights to imputed data for analysis.
}
\examples{
{
  # TO-DO
}
}
