% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_wgts.R
\name{append_attrition_weights_to_df}
\alias{append_attrition_weights_to_df}
\title{Add Attrition Weights to Raw Data}
\usage{
append_attrition_weights_to_df(
  data,
  country = NULL,
  composite.wgt.name = NULL,
  attr.wgt.name = NULL,
  wgt = NULL,
  strata = NULL,
  psu = NULL
)
}
\arguments{
\item{data}{a pre-cleaned and mostly ready for analysis data.frame that will be used to construct
attrition weights.}

\item{composite.wgt.name}{(default NULL) name of composite weight in saved data}

\item{attr.wgt.name}{(default NULL) name of attrition weight in saved data}

\item{attr.wgt.file}{(default NULL), a file name to get the stored attrition weights}
}
\value{
a dataset with attrition weights appended
}
\description{
TO-DO
}
\details{
Adds estimated attrition weights to raw data to help summarize data.
}
\examples{
{
  # TO-DO
}
}
