% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_build_tbl_fig.R
\name{build_tbl_1}
\alias{build_tbl_1}
\alias{build_tbl_2}
\alias{build_tbl_3}
\alias{build_tbl_sample_by_x}
\alias{build_tbl_outcome_by_x}
\alias{build_tbl_variable_by_x}
\alias{build_tbl_attr_model}
\alias{build_tbl_country_point_estimates}
\alias{build_tbl_evalues}
\alias{build_tbl_outcomewide}
\alias{build_tbl_predictorwide}
\alias{build_tbl_predictorwide_evalues}
\alias{build_tbl_predictorwide_pnt_est_wide}
\alias{build_tbl_outcomes_exta_wide}
\alias{build_tbl_sample_extra_wide}
\alias{build_tbl_pca_summary}
\alias{build_fig_1}
\alias{gfs_supp_forest_plot}
\alias{generate_docx_normal_portrait}
\alias{generate_docx_normal_landscape}
\alias{generate_docx_wide_landscape}
\alias{generate_docx_fig}
\title{Internal Build Table Functions}
\usage{
build_tbl_1(params, font.name = "Open Sans", font.size = 10)

build_tbl_2(params, font.name = "Open Sans", font.size = 10)

build_tbl_3(params, font.name = "Open Sans", font.size = 10)

build_tbl_sample_by_x(params, font.name = "Open Sans", font.size = 10)

build_tbl_outcome_by_x(params, font.name = "Open Sans", font.size = 10)

build_tbl_variable_by_x(params, font.name = "Open Sans", font.size = 10)

build_tbl_attr_model(params, font.name = "Open Sans", font.size = 10)

build_tbl_country_point_estimates(
  params,
  font.name = "Open Sans",
  font.size = 9
)

build_tbl_evalues(params, font.name = "Open Sans", font.size = 10)

build_tbl_outcomewide(params, font.name = "Open Sans", font.size = 10)

build_tbl_predictorwide(params, font.name = "Open Sans", font.size = 10)

build_tbl_predictorwide_evalues(
  params,
  font.name = "Open Sans",
  font.size = 10
)

build_tbl_predictorwide_pnt_est_wide(
  params,
  font.name = "Open Sans",
  font.size = 9
)

build_tbl_outcomes_exta_wide(params, font.name = "Open Sans", font.size = 10)

build_tbl_sample_extra_wide(params, font.name = "Open Sans", font.size = 10)

build_tbl_pca_summary(params, font.name = "Open Sans", font.size = 10)

build_fig_1(
  params,
  forest.plots.inc.est = FALSE,
  font.name = "Open Sans",
  font.size = 10
)

gfs_supp_forest_plot(params, plot.type = "combined", ...)

generate_docx_normal_portrait(cache.file, print.file)

generate_docx_normal_landscape(cache.file, print.file)

generate_docx_wide_landscape(cache.file, print.file)

generate_docx_fig(cache.file, print.file, fig.file, orient = "p", h = 6, w = 5)
}
\arguments{
\item{params}{a list of parameters that were originally passed as parameters in the .Rmd files. Kept for legacy and to reduce need to rewrite code.}

\item{font.name}{"Open Sans"}

\item{font.size}{10}

\item{...}{additional arguments as needed}

\item{fit}{fitted results object from the metafor package (rma)}

\item{better.name}{a manually supplied name for the focal predictor}

\item{p.title}{a string passed to title of the forest plot}

\item{p.subtitle}{a string passed to subtitle of the forest plot.}
}
\value{
a ggplot object
}
\description{
Generate the caches table object for printing: Main text summary table
}
\examples{
# TO-DO

}
