% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_variance}
\alias{check_variance}
\title{Check Variable Variance}
\usage{
check_variance(x, data)
}
\arguments{
\item{x}{(character) a variable name in data}

\item{data}{(data.frame)}
}
\value{
a logical (TRUE/FALSE)
}
\description{
checks the variances of the predictor variable, identifies whether the variable "x" have any
variance (works w/ any variable type, continuous, binary, nominal,...).
}
\examples{
{
  df <- data.frame(A = c(1, 1, 0, 0), B = c("B", "B", "B", "B"))
  check_variance("A", df)
  check_variance("B", df)
}
}
