% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_meta_analysis_utils.R
\name{compute_calibrated}
\alias{compute_calibrated}
\alias{proportion_meaningful}
\alias{get_meta_ci}
\alias{compute_global_pvalue}
\alias{add_pop_wgts}
\alias{construct_meta_input_from_saved_results}
\title{Meta-Analysis Utility Functions}
\usage{
compute_calibrated(fit)

proportion_meaningful(
  x,
  q = 0,
  above = TRUE,
  method = "empirical",
  theta = 0,
  tau = 1
)

get_meta_ci(fit, type = "Q")

compute_global_pvalue(x, p.min = 1e-16)

add_pop_wgts(df)

construct_meta_input_from_saved_results(res.dir, outcomes, predictors)
}
\arguments{
\item{fit}{the resulting object from fitting the metafor::rma(.) function.}

\item{x}{vector of calibrated effect sizes}

\item{q}{a numerical values (default = 0) that defines the threshold of meaningful effect size}

\item{above}{a logical indicator of whether to come the proportion above or below the threshold}

\item{method}{a character string defining how the proportion should be computed (default is "normal" but can be "empirical")}

\item{theta}{the population average effect size}

\item{tau}{the population standard deviation}
}
\value{
depends on function...
}
\description{
A set of utility functions used in the post-processing of the meta-analysis results.
}
\examples{
# TO-DO
}
