% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_wgts.R
\name{create_attr_wgts}
\alias{create_attr_wgts}
\title{Estimate and Save Attrition Weights}
\usage{
create_attr_wgts(
  fit,
  obs.id.var = NULL,
  stabilized = TRUE,
  wgt.trim.quantile = 0.99,
  wgt = "ANNUAL_WEIGHT_R2",
  strata = "STRATA",
  psu = "PSU",
  composite.wgt.name = NULL,
  attr.wgt.name = NULL,
  ...
)
}
\arguments{
\item{fit}{(model object) that was fit to estimate attrition weights}

\item{stabilized}{(logical) default to TRUE, where the estimated attrition weight is the divisor and the numerator is the baseline rate of attrition (if FALSE, then the numerator is fixed to 1)}

\item{wgt.trim.quantile}{(numeric) default to 0.99, quantile of the distribution of weights to trim to}

\item{wgt}{(character) name of variable containing weights to use as "base-weights"}

\item{strata}{(character) name of variable containing strata ids}

\item{psu}{(character) name of variable containing primary sampling unit ids}

\item{...}{other arguments passed to svyglm or glmrob functions}
}
\value{
a dataset with attrition weights appended
}
\description{
The default set of predictors: "ANNUAL_WEIGHT_R2", "MODE_ANNUAL_W1", "AGE_W1", "GENDER_W1",
"EDUCATION_3_W1", "EMPLOYMENT_W1", "MARITAL_STATUS_W1", "RACE_PLURALITY_W1".
}
\details{
Estimate attrition weights to data.frame
}
\examples{
{
  # TO-DO
}
}
