% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_if_valid_comb}
\alias{.check_if_valid_comb}
\title{identify if predictor/outcome/country combination is valid}
\usage{
.check_if_valid_comb(
  cur.country,
  your.outcome,
  your.pred,
  covariates,
  country.subset = NULL
)
}
\description{
identify if predictor/outcome/country combination is valid
}
\keyword{internal}
