% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_data}
\alias{.get_data}
\title{Function to get data from saved files}
\usage{
.get_data(
  file,
  data.dir,
  wgt = ANNUAL_WEIGHT_R2,
  psu = PSU,
  strata = STRATA,
  imp.strata = COUNTRY,
  direct.subset = NULL,
  domain.subset = NULL,
  pc.rule = "omit",
  pca.variables = NULL
)
}
\description{
Function to get data from saved files
}
\keyword{internal}
