% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_get_variables.R
\name{get_demo_childhood_variables}
\alias{get_demo_childhood_variables}
\title{Get variable codes for demographic and childhood variables}
\usage{
get_demo_childhood_variables(...)
}
\arguments{
\item{...}{argument not used}
}
\value{
a character string
}
\description{
A simple function to get a list of demographic or childhood predictor variables.
}
\examples{
{
  get_demo_childhood_variables()
}
}
