% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_pca.R
\name{get_eigenvalues}
\alias{get_eigenvalues}
\title{Get Eigenvalues}
\usage{
get_eigenvalues(data, var, cor = FALSE)
}
\arguments{
\item{data}{a data.frame or survey.design object}

\item{var}{a character vector of variable names}

\item{cor}{a logical of whether to get eigenvalues based on correlation matrix instead of covariance matrix}
}
\value{
a vector of eigenvalues
}
\description{
When data is a survey.design object, the eigenvalues are estimated based on the weighted
covariance matrix estimated with the stats package function cov.wt(.) using the design
probabilities.
}
\details{
Extacts the estimated eigenvalues for summary later.
}
\examples{
{
  # to-do
}
}
