% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_variables_2w_test.R
\name{get_missing_codes}
\alias{get_missing_codes}
\title{Get Outcome Missingness Codes}
\usage{
get_missing_codes(var)
}
\arguments{
\item{var}{a character string (e.g., 'HAPPY_W1')}
}
\value{
a character string
}
\description{
A relatively simple switch function to get the missingness codes (-98, 98, 99, etc.)
}
\examples{
{
  get_missing_codes("HAPPY_W1")
}
}
