% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_wave_3_utils.R
\name{get_tbl_row_vec}
\alias{get_tbl_row_vec}
\title{utility function to get the prespecified set of variables used in the rows of the tables.
blank',}
\usage{
get_tbl_row_vec(study = "exposurewide", filetype = "main")
}
\description{
utility function to get the prespecified set of variables used in the rows of the tables.
blank',
}
