% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_childhood_variables.R
\name{get_variable_codes}
\alias{get_variable_codes}
\title{Get variable codes other}
\usage{
get_variable_codes(...)
}
\arguments{
\item{...}{argument not used}
}
\value{
a character string
}
\description{
A simple function to get a list of variable codes used throughout the package
}
\examples{
{
  get_variable_codes()
}
}
