% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_wave_3_utils.R
\name{get_variable_group_labels}
\alias{get_variable_group_labels}
\title{Attached outcome/exposure label group to data.frame}
\usage{
get_variable_group_labels(x)
}
\description{
Attached outcome/exposure label group to data.frame
}
