% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variable_value_switch_funtions.R
\name{get_wave_flag}
\alias{get_wave_flag}
\title{Get Wave by Item Administered Falg}
\usage{
get_wave_flag(var, wave = NULL)
}
\arguments{
\item{var}{a character string of what variable to check (e.g., "HAPPY")}

\item{wave}{an optional character string to check against. Must be either one of NULL, 'W1', 'WR1', 'W2'.}
}
\value{
a character vector or logical
}
\description{
A switch function to identify what wave(s) a variable was administered OR to check where a variable was administered in a particular wave.
}
\examples{
{
 # TO-DO
}
}
