% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variable_value_switch_funtions.R
\name{gfs_add_variable_labels}
\alias{gfs_add_variable_labels}
\title{Get outcome name (better)}
\usage{
gfs_add_variable_labels(df, vars, ...)
}
\arguments{
\item{df}{data.frame with column (var)}

\item{...}{other parameters passes to get_outcome_better_name(.)}

\item{var}{a character string (e.g., 'HAPPY_W1')}
}
\value{
the df with label attribute updated
}
\description{
A relatively simple switch function to get a better name for each variable
}
\examples{
{
  # TO-DO
}
}
