% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_ft_themes.R
\name{format_flex_table}
\alias{format_flex_table}
\alias{theme_meta_outcome_wide}
\alias{theme_meta_evalues}
\alias{theme_meta_predictor_wide}
\title{GFS Flextable Themes}
\usage{
format_flex_table(xtb, pg.width = 6.5)

theme_meta_outcome_wide(xtb, pg.width = 9, .ncol = 12)

theme_meta_evalues(xtb, pg.width = 6.5, .ncol = 6)

theme_meta_predictor_wide(xtb, pg.width = 9, .ncol = 10)
}
\arguments{
\item{xtb}{a flextable object}

\item{pg.width}{width of page to adjust column widths to.}
}
\value{
a formatted flextable
}
\description{
TODO
}
\details{
Possible themes with various defaults to help with formatting tables.
}
\examples{
{
  ## TODO
}
}
