% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_coordinated_tables.R
\name{gfs_generate_supplemental_docs}
\alias{gfs_generate_supplemental_docs}
\title{Construct GFS Online Supplement results}
\usage{
gfs_generate_supplemental_docs(
  df.raw = NULL,
  meta.wopc = NULL,
  meta.wpc = NULL,
  coun.wopc = NULL,
  coun.wpc = NULL,
  coun.fit.pca = NULL,
  focal.predictor = NULL,
  focal.better.name = "Focal Predictor",
  focal.predictor.reference.value = "estimated population mean of focal predictor",
  p.bonferroni = NULL,
  baseline.pred = NULL,
  outcome.vec = NULL,
  mylabels = NULL,
  res.dir = "results",
  included.countries = NULL,
  what = "both"
)
}
\arguments{
\item{df.raw}{a data.frame containing the raw data with values coded as labels}

\item{meta.wopc}{a nested data.frame of results from the meta-analysis function (see gfs_meta_analysis(.)) of the meta-analytic results that did NOT include the principal components in the country-specific regression analyses}

\item{meta.wpc}{a nested data.frame of results from the meta-analysis function (see gfs_meta_analysis(.)) of the meta-analytic results that did INCLUDED the principal components in the country-specific regression analyses}

\item{coun.wopc}{a data.frame of country-specific results (e.g., input of meta-analytic) function (see gfs_meta_analysis(.)) of the results that did NOT include the principal components in the country-specific regression analyses}

\item{coun.wpc}{a data.frame of country-specific results (e.g., input of meta-analytic) function (see gfs_meta_analysis(.)) of the results that did INCLUDED the principal components in the country-specific regression analyses}

\item{focal.predictor}{a character defining the focal predictor}

\item{focal.better.name}{a character that is used as the printed name in tables/captions to denote the focal predictor}

\item{focal.predictor.reference.value}{(character) describing the baseline/reference group for the focal predictor.}

\item{p.bonferroni}{a number (e.g., 0.007), is internally determined based on number of rows in coun.wopc}

\item{baseline.pred}{a vector of characters defining which baseline characteristics were used as control variables. This can be used to force the inclusion some variable into the main text summary table.}

\item{outcome.vec}{a character vector of outcomes names (e.g., "HAPPY_W2") that are to be printed in the main text meta-analytic summary table. Name MUST be included in the coun.wopc (coun.wpc) nested data.frames column (OUTCOME0), otherwise the variable won't be printed.}

\item{mylabels}{an optional character vector that will be printed out in specific rows of tables 2/3 depending on the specification pf outcome.vec}

\item{res.dir}{(character) defaults to "results", and will be created if needed to story results document}

\item{included.countries}{a character vector of which countries to include in output -- defaults to all.}

\item{what}{(character) options include: "both", "S1", "S2"}

\item{...}{other arguments as needed}
}
\value{
a word document saved to the current 'results/' directory
}
\description{
Generates two online supplements.

Supplement 1:

(1) Summary statistics of OUTCOMES by wave (raw data)
(2) Country-specific results
   - Summary statistics by wave for demographics (similar to main text Table 1)
   - Summary statistics by wave for outcomes (similar to Table S1)
   - Summary statistics of principal components by outcome (# retained, % prop explained, cumsum % prop explained)
   - Outcome-wide results (similar to main text Table 2)
   - Outcome-wide E-values (similar to main text Table 3)

Supplement 2:

(1) Summary statistics of demographics by country (raw data)
(2) Summary statistics of OUTCOMES by country (raw data)
(3) Forest plots of all effects
    - Model 1 (No PCs)
    - Model 2 (w/ PCs)
}
\details{
Generated two word documents containing the supplemental results for the outcome wide results for the GFS coordinated analyses.
}
\examples{
{
  # TO-DO
}
}
