% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{gfs_get_labelled_raw_data}
\alias{gfs_get_labelled_raw_data}
\title{Estimate and Save Attrition Weights}
\usage{
gfs_get_labelled_raw_data(file, list.composites = NULL, ...)
}
\arguments{
\item{file}{name of file to get data from}

\item{list.composites}{a lit of lists to construct composite variables}

\item{...}{other arguments}
}
\value{
a dataset with attrition weights appended
}
\description{
TO-DO
}
\details{
Estimate (save fitted model) and append attrition weights to data.frame
}
\examples{
{
  # TO-DO
}
}
