% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_meta_analysis.R
\name{gfs_meta_analysis}
\alias{gfs_meta_analysis}
\title{GFS Wrapper for Meta-Analysis}
\usage{
gfs_meta_analysis(
  meta.input,
  yi = std.est,
  sei = std.se,
  grpi = Country,
  dfi = df,
  pvalue = p.value,
  estimator = "PM",
  interval.method = "empirical",
  reliability.vec = c(0.4, 0.55, 0.7),
  var.name = NULL,
  ci.alpha = 0.05,
  better.name = NULL,
  p.subtitle = NULL
)
}
\arguments{
\item{meta.input}{a nested data.frame where `data` is a column}

\item{yi}{a name of the variable containing the effect sizes to be meta-analyzed (e.g., as.name("Est"))}

\item{sei}{a name of the variable containing the standard error of the effect size (e.g., as.name("SE"))}

\item{grpi}{the name of the variable used to identify each "study" or group, (defaults to 'as.name("Country")')}

\item{estimator}{(default = "PM") other options include those available in the metafor package}

\item{interval.method}{(default = "empirical) how the proportion of effect size metric is computed}

\item{ci.alpha}{(0.05) type-1 error rate (alpha) for confidence intervals}

\item{...}{additional arguments}
}
\value{
a list of meta-analysis results from the metafor package
}
\description{
Conducted meta-analysis (random effects or population weighted) for a set of effect of interest.
}
\examples{
# TO-DO
}
