% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_meta_forest_plot.R
\name{gfs_meta_forest_plot}
\alias{gfs_meta_forest_plot}
\title{GFS Formatted Forest Plots}
\usage{
gfs_meta_forest_plot(
  fit,
  better.name = NULL,
  p.subtitle = "GFS Forest Plot",
  p.title = NULL,
  ...
)
}
\arguments{
\item{fit}{fitted results object from the metafor package (rma)}

\item{better.name}{a manually supplied name for the focal predictor}

\item{p.subtitle}{a string passed to subtitle of the forest plot.}

\item{p.title}{a string passed to title of the forest plot}

\item{...}{additional arguments as needed}
}
\value{
a ggplot object
}
\description{
Forest plots constructed from GFS meta analysis results for use in manuscripts.
}
\examples{
# TO-DO

}
