% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_pool_estimates.R
\name{gfs_pool_estimates}
\alias{gfs_pool_estimates}
\title{Pool Results across Imputations}
\usage{
gfs_pool_estimates(data)
}
\arguments{
\item{data}{a long.data.frame with at least a .imp column and a fit.tidy}
}
\value{
a one-row data.frame with columns
* nimp
* estimate.pooled
* se.pooled
* df.approx
* t.statistic
* f.statistic
* ci.low
* ci.up
* p.value
see below for additional information about these statistics
}
\description{
TO-DO
}
\details{
Pool results across imputed datasets using Rubin's Rules.
}
