% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_coordinate_analysis_function.R
\name{gfs_run_regression_single_outcome}
\alias{gfs_run_regression_single_outcome}
\title{GFS Focal Predictor Analysis}
\usage{
gfs_run_regression_single_outcome(
  data = NULL,
  your.pred = NULL,
  your.outcome = NULL,
  covariates = NULL,
  contemporaneous.exposures = NULL,
  wgt = ANNUAL_WEIGHT_R2,
  psu = PSU,
  strata = STRATA,
  standardize = TRUE,
  pc.cutoff = 7,
  pc.rule = "omit",
  force.linear = FALSE,
  force.binary = FALSE,
  robust.huberM = FALSE,
  robust.tune = 1,
  res.dir = NULL,
  list.composites = NULL,
  subpopulation = NULL,
  compute.vif = FALSE,
  appnd.txt.to.filename = "",
  .return.all = FALSE,
  ...
)
}
\arguments{
\item{data}{long data.frame must have COUNTRY and .imp columns (default: df.imp.long)}

\item{your.pred}{a character string defining focal predictor (e.g., "PHYSICAL_HLTH_Y1")}

\item{your.outcome}{a character string defining outcome variable (e.g., "HAPPY_Y2")}

\item{covariates}{a character vector defining core set of covariates (default: NULL)}

\item{contemporaneous.exposures}{a character vector defining set of contemporaneous exposures that are input into PCA (default: NULL)}

\item{standardize}{a logical (default: TRUE) of whether to standardized the predictor and outcome}

\item{pc.cutoff}{a numeric value, can either be a fixed whole number (e.g., keep 7 PC in all countries) OR a proportion (e.g., 0.50)}

\item{pc.rule}{principal components analysis use rule (default: "omit")}

\item{force.linear}{a logical of whether to force a linear model (default: FALSE)}

\item{force.binary}{a logical of whether to force a Poisson model (default: FALSE)}

\item{robust.huberM}{a  logical of whether to use a robust variant of the linear regression model (default: FALSE), see below for additional details.}

\item{robust.tune}{a numeric value defining the tuning parameter for the robust.huberM option.}

\item{res.dir}{a character string defining directory to save results to.}

\item{subpopulation}{(optional list) of length up to three defining the subdomain to the analyzed (see examples for how argument is structured)}

\item{compute.vif}{(optional) a logical of whether to compute the VIF. *warning* this significantly increases the time needed to get results.}

\item{appnd.txt.to.filename}{(optional) character string to help differentiate saved results file (default "")}

\item{...}{other arguments passed to svyglm or glmrob functions}
}
\value{
a data.frame that contains the meta-analysis input results
}
\description{
The `pc.rule` argument determines how should the number of PCs is decided:
exclude PCs from analysis (default: "omit");
a fixed number ("constant");
minimunm total proportion of variance explained ("mintotal"); or
minimum proportion explain by individual PCs ("mincomp").
For "constant", `pc.cutoff` must of a number between 1 and length(contemporaneous.exposures) - 1.
For "mintotal", `pc.cutoff` must be strictly between 0 and 1, e.g. (0.70).
For "mincomp", `pc.cutoff` must be strictly between 0 and 1, e.g. (0.01).

When standardize is TRUE, standardize the outcome and predictor using the survey-adjusted mean and variance.

When force.linear is TRUE, forces the estimation of a linear model even for binary/Likert outcomes.
I think this is sometimes called as a "linear probability model."

When robust.huberM is TRUE, for (approximately) continuous outcomes, you have the option of
alternatively using a "robust m-estimator" with Huber style robustness weights in addition to
the complex sampling design adjustments. It's unknown whether this makes a meaningful difference,
but preliminary testing suggests small differences in point estimates but sometimes dramatic
changes to standard errors for reasons that are unclear to me. Could be due to a strange
interaction of robustness weights, attrition weights, and post-stratified sampling weights.
}
\details{
Run regression analyses for specific focal predictor using the core GFS group analysis pathway.
}
\examples{
{
  # TO-DO
}
}
