% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_svyglm_mi.R
\name{gfs_svyglm_mi}
\alias{gfs_svyglm_mi}
\title{GFS Wrapper for generalized linear models with multiple imputation}
\usage{
gfs_svyglm_mi(
  fx = NULL,
  data.dir = NULL,
  wgt = ANNUAL_WEIGHT_R2,
  psu = PSU,
  strata = STRATA,
  force.linear = FALSE,
  force.binary = FALSE,
  robust.huberM = FALSE,
  robust.tune = 1,
  res.dir = "results",
  direct.subset = NULL,
  domain.subset = NULL,
  country.subset = NULL,
  family = NULL,
  appnd.txt.to.filename = "",
  save.all = FALSE,
  ...
)
}
\arguments{
\item{fx}{(NULL) a formula argument that can be used instead of your.pred/your.outcome to specify the set of predictors and outcome}

\item{data.dir}{a character string defining where data are located}

\item{force.linear}{a logical of whether to force a linear model (default: FALSE)}

\item{force.binary}{a logical of whether to force a Poisson model (default: FALSE)}

\item{robust.huberM}{a  logical of whether to use a robust variant of the linear regression model (default: FALSE), see below for additional details.}

\item{robust.tune}{a numeric value defining the tuning parameter for the robust.huberM option.}

\item{res.dir}{a character string defining directory to save results to.}

\item{direct.subset}{a quoted subset statement to directly subset the sample}

\item{domain.subset}{a quoted subset statement to run domain estimation}

\item{country.subset}{a string vector of countries included in analysis}

\item{appnd.txt.to.filename}{(optional) character string to help differentiate saved results file (default "")}

\item{save.all}{(FALSE) logical of whether to save the fitted regression models from each imputed dataset (this is a double-checking feature and should generally be set to FALSE unless you plan to probe the individual regression models--see the methods paper for Wave 2 for example of it's used)}

\item{...}{other arguments passed to svyglm or glmrob functions}
}
\value{
a data.frame that contains the meta-analysis input results
}
\description{
When force.linear is TRUE, forces the estimation of a linear model even for binary/Likert outcomes.
I think this is sometimes called as a "linear probability model."

When robust.huberM is TRUE, for (approximately) continuous outcomes, you have the option of
alternatively using a "robust m-estimator" with Huber style robustness weights in addition to
the complex sampling design adjustments. It's unknown whether this makes a meaningful difference,
but preliminary testing suggests small differences in point estimates but sometimes dramatic
changes to standard errors for reasons that are unclear to me. Could be due to a strange
interaction of robustness weights, attrition weights, and post-stratified sampling weights.
}
\details{
Run GLMs with the folder structure for memory ease and pooling across imputed datasets. Cannot incorporate principle compoents into the analysis; if that is desired using the gfs_run_regression_single_outcome(.) function.
}
\examples{
{
  # TO-DO
}
}
