% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_wave_3_tables.R
\name{gfs_wave_3_build_tbl_1}
\alias{gfs_wave_3_build_tbl_1}
\title{Specific table functions
Generate the caches table object for printing: Main text summary table}
\usage{
gfs_wave_3_build_tbl_1(params, font.name = "Open Sans", font.size = 10)
}
\arguments{
\item{params}{a list of parameters that were originally passed as parameters in the .Rmd files. Kept for legacy and to reduce need to rewrite code.}

\item{font.name}{"Open Sans"}

\item{font.size}{10}
}
\description{
Specific table functions
Generate the caches table object for printing: Main text summary table
}
\keyword{internal}
