% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_wave_3_tables.R
\name{gfs_wave_3_generate_main_doc}
\alias{gfs_wave_3_generate_main_doc}
\title{Construct GFS Main text results (Wave 3)}
\usage{
gfs_wave_3_generate_main_doc(
  df.raw = NULL,
  focal.variable = NULL,
  focal.better.name = "Focal Variable",
  focal.variable.reference.value = "0",
  digits = 2,
  control = list(study = "exposurewide", filetype = "main")
)
}
\arguments{
\item{df.raw}{a data.frame containing the raw data with values coded as labels}

\item{focal.better.name}{a character that is used as the printed name in tables/captions to denote the focal predictor}

\item{focal.variable.reference.value}{(character) describing the baseline/reference group for the focal predictor.}

\item{dir.meta}{subdirectory where primary results are stored ('results-primary')}

\item{file.wopc}{file containing the nested-data.frame(.) containing the meta-analyzed results with principal components included}

\item{p.bonferroni}{a number (e.g., 0.00081), is internally determined based on length of tbl.row.vec if not provided}

\item{baseline.pred}{a vector of characters defining which baseline characteristics were used as control variables. This can be used to force the inclusion some variable into the main text summary table.}

\item{tbl.row.vec}{a character vector of outcomes names (e.g., "HAPPY_Y2") that are to be printed in the main text meta-analytic summary table. Name MUST be included in the meta.wopc (meta.wpc) nested data.frames column (OUTCOME0), otherwise the variable won't be printed.}

\item{mylabels}{an optional character vector that will be printed out in specific rows of tables 2/3 depending on the specification pf tbl.row.vec}

\item{res.dir}{(character) defaults to "results", and will be created if needed to story results document}

\item{...}{other arguments as needed}
}
\value{
a word document saved to the current 'results/' directory
}
\description{
TO-DO
}
\details{
Generated a word document containing the results for the meta-analytic outcome wide results for the GFS coordinated analyses.
}
\examples{
{
  # TO-DO
}
}
