% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Load Packages}
\usage{
load_packages(...)
}
\arguments{
\item{...}{no direct parameters}
}
\value{
nothing (only background global environment changes)
}
\description{
Function to load the required packages into the environment
}
\examples{
{
  load_packages()
}
}
