% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quickpred2}
\alias{quickpred2}
\title{Create Predictor Matrix (modified)}
\usage{
quickpred2(
  data,
  mincor = 0.1,
  minpuc = 0,
  include = "",
  exclude = "",
  method = "pearson",
  maxcor = 0.71
)
}
\arguments{
\item{data}{Matrix or data frame with incomplete data.}

\item{mincor}{A scalar, numeric vector (of size \code{ncol(data))} or numeric
matrix (square, of size \code{ncol(data)} specifying the minimum
threshold(s) against which the absolute correlation in the data is compared.}

\item{minpuc}{A scalar, vector (of size \code{ncol(data))} or matrix (square,
of size \code{ncol(data)} specifying the minimum threshold(s) for the
proportion of usable cases.}

\item{include}{A string or a vector of strings containing one or more
variable names from \code{names(data)}. Variables specified are always
included as a predictor.}

\item{exclude}{A string or a vector of strings containing one or more
variable names from \code{names(data)}. Variables specified are always
excluded as a predictor.}

\item{method}{A string specifying the type of correlation. Use
\code{'pearson'} (default), \code{'kendall'} or \code{'spearman'}. Can be
abbreviated.}

\item{maxcor}{A scalar, numeric vector (of size \code{ncol(data))} or numeric
matrix (square, of size \code{ncol(data)} specifying the MAXIMUM
threshold(s) against which the absolute correlation in the data is compared.
this is used to help troubleshoot and resolve non-invertible predictor matrices}
}
\value{
a predictor matrix for use in mice
}
\description{
Function copied and modified from mice pacakge (https://github.com/amices/mice/blob/master/R/quickpred.R)
}
