% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoding_imputed_data_2w_test.R
\name{recode_imputed_data}
\alias{recode_imputed_data}
\title{Reformat Imputed Data}
\usage{
recode_imputed_data(
  df.imp = NULL,
  list.default = NULL,
  list.composites = NULL,
  list.manual = NULL,
  ...
)
}
\arguments{
\item{df.imp}{a nested data.frame with results from run_impute_data(.).}

\item{list.default}{(required) a names list with DEMOGRAPHICS.CHILDHOOD.PRED.VEC, OUTCOME.VE,
FOCAL_PREDICTOR, USE_DEFAULT, FORCE_BINARY, VALUES_DEFINING_UPPER_CATEGORY, VALUES_DEFINING_LOWER_CATEGORY, FORCE_CONTINUOUS}

\item{list.composites}{(optional) a named list with elements LIST.OUTCOME.COMPOSITES, LIST.COMPOSITE.COMBINE.METHOD and COMPOSITE.VEC.}

\item{list.manual}{(optional) a list like list.default (to-do)}

\item{...}{other arguments passed to mice}
}
\value{
a long data.frame for use in analyses
}
\description{
TO-DO
}
\details{
Reformated the imputed (nested) data for use in coordinated regression analyses.
}
\examples{
{
  # none
}
}
