% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_variables_2w_test.R
\name{recode_labels}
\alias{recode_labels}
\title{Recode Variable Labels}
\usage{
recode_labels(x, var)
}
\arguments{
\item{x}{a vector}

\item{var}{a character string (e.g., 'HAPPY_W1')}
}
\value{
a vector
}
\description{
A relatively simple switch function to transform variables to from the numeric values
to the labels associated with those levels.
}
\examples{
{
  # TO-DO
}
}
