% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variable_value_switch_funtions.R
\name{recode_race_to_plurality}
\alias{recode_race_to_plurality}
\title{Recode Race to Binary Plurality}
\usage{
recode_race_to_plurality(x, country)
}
\arguments{
\item{x}{a vector}

\item{country}{a character string (e.g., 'COUNTRY')}
}
\value{
a vector
}
\description{
A case when to recode RACE into a binary race plurality flag of whether a participant is in the majority race for their country.
}
\examples{
{
  # to-do
}
}
