% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variable_value_switch_funtions.R
\name{recode_to_numeric}
\alias{recode_to_numeric}
\title{Recode Variable to Numeric}
\usage{
recode_to_numeric(x, var, is.sum = FALSE)
}
\arguments{
\item{x}{a vector}

\item{var}{a character string (e.g., 'HAPPY_Y1')}

\item{is.sum}{(default FALSE) a logical determining whether to change how a variable is treated.
Currently only applies to certain composite variables.}
}
\value{
a vector
}
\description{
A more complex switch function that is highly dependent on the variable and needs to be reviewed
for accuracy.
}
\examples{
{
  # to-do
}
}
