% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variable_value_switch_funtions.R
\name{recode_to_type}
\alias{recode_to_type}
\title{Recode Variable to Type}
\usage{
recode_to_type(x, var)
}
\arguments{
\item{x}{a vector}

\item{var}{a character string (e.g., 'HAPPY_W1')}
}
\value{
a vector
}
\description{
A relatively simple switch function to transform variables to factors, numeric, etc.
}
\examples{
{
  # to-do
}
}
