% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_reliability_sensitivity.R
\name{reliability_corrected_estimates}
\alias{reliability_corrected_estimates}
\title{Reliability Sensitivity Analysis}
\usage{
reliability_corrected_estimates(theta, se, crit, lambda)
}
\arguments{
\item{theta}{baseline parameter value}

\item{se}{standard error of the estimate}

\item{crit}{critical value of t-distribution (e.g., qt(1-ci.alpha/2, df)=qt(0.975,98))}

\item{lambda}{possibly a vector of average reliability coefficients for predictor/outcome combination}
}
\value{
a data.frame
}
\description{
The reliability sensitivity analysis is based on Spearman attenuate correction for measures of association (Spearman, 1904). The correction aims to provide a "corrected" estimate of association after accounting for the attentuation that occurs due to uncorrelated measurement error.
The correction is
$$b_{corrected} = \frac{b}{\lambda}$$
}
\details{
Calculated reliability corrected estimates of regression coefficients/
}
\examples{
reliability_corrected_estimates(0.10, 0.02, 1.96, c(0.40, 0.55, 0.70))
}
