% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variable_value_switch_funtions.R
\name{reorder_levels}
\alias{reorder_levels}
\title{Recode Variable Based on Levels}
\usage{
reorder_levels(x, var)
}
\arguments{
\item{x}{a vector}

\item{var}{a character string (e.g., 'HAPPY_W1')}
}
\value{
a vector
}
\description{
A relatively simple switch function to transform variables by either reordering levels (reverse-
coding or otherwise).
}
\examples{
{
  # to-do
}
}
