% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rglobalflourishing-package.R
\docType{package}
\name{Rglobalflourishing-package}
\alias{Rglobalflourishing}
\alias{Rglobalflourishing-package}
\title{Rglobalflourishing: Combined Analysis Workflow for the Global Flourishing Study Core Team}
\description{
A structured set of analyses for analyzing data from the Global Flouriishing Study. The package is primarily intended for use by the core GFS research team (HFH, Baylor RSI, among others). The package has additional functionality that extends beyond the core set of analyses to help facilitate the use ease of use of the GFS data such as relative simple regression analyses and summary statistics.
}
\author{
\strong{Maintainer}: Noah Padgett \email{npadgett@hsph.harvard.edu}

Authors:
\itemize{
  \item Chris Felton \email{chris_felton@baylor.edu} [contributor]
  \item Byron Johnson \email{byron_johnson@baylor.edu} [funder]
  \item Tyler VanderWeele \email{tvanderw@hsph.harvard.edu} [funder]
}

}
\keyword{internal}
