% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_wgts.R
\name{run_attrition_model}
\alias{run_attrition_model}
\title{Wrapper Function Attrition Weights}
\usage{
run_attrition_model(data, ...)
}
\arguments{
\item{data}{a pre-cleaned and mostly ready for analysis data.frame that will be used to construct
attrition weights.}

\item{...}{additional arguments passed to append_attrition_wgts(.)}
}
\value{
a dataset with attrition weights appended
}
\description{
Wrapper function to use the data.set specific attrition weights function.
}
\examples{
{
  # TO-DO
}
}
