% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_wgts.R
\name{run_attrition_model_by_country}
\alias{run_attrition_model_by_country}
\title{Wrapper Function Attrition Weights}
\usage{
run_attrition_model_by_country(
  data.dir,
  pooled.wgt = NULL,
  composite.wgt.name = NULL,
  ...
)
}
\arguments{
\item{...}{additional arguments passed to ....}
}
\value{
a dataset with attrition weights appended
}
\description{
TO-DO
}
\details{
Wrapper function to use the data.set specific attrition weights function.
}
\examples{
{
  # TO-DO
}
}
