% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{svy_scale}
\alias{svy_scale}
\title{Standardized with Complex Survey Adjusted Mean and Variance}
\usage{
svy_scale(x, wgt = NULL, psu = NULL, strata = NULL, design = NULL)
}
\arguments{
\item{x}{(character) a variable name in data}

\item{wgt}{(character) variable name of weights}

\item{psu}{(character) variable name of PSUs}

\item{strata}{(character) variable name of stratum ids}

\item{design}{(survey.design or survey.design2 object)}
}
\value{
a vector of standardized scores
}
\description{
Standardizes a variable x based on the weights and design. A better replacement for scale(.).
}
\examples{
{
  library(survey)
  # TODO
}
}
