% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variable_value_switch_funtions.R
\name{get_outcome_better_name}
\alias{get_outcome_better_name}
\title{Get outcome name (better)}
\usage{
get_outcome_better_name(
  var,
  include.name = TRUE,
  include.wave = FALSE,
  include.fid = FALSE
)
}
\arguments{
\item{var}{a character string (e.g., 'HAPPY_Y1')}

\item{include.wave}{(logical) whether to include wave identifier (Wave 1) in returned string, not applicable when include.name is true}

\item{include.fid}{(logical) whether to include a flag (a-e) for designating which flourishing index domain each an item belongs to}

\item{include.label}{(logical) whether to include variable name (var) in returned string}
}
\value{
a character string
}
\description{
A relatively simple switch function to get a better name for each variable
}
\examples{
{
  get_outcome_better_name("HAPPY_Y1")
  get_outcome_better_name("HAPPY_Y1", include.name = FALSE)
  get_outcome_better_name("HAPPY_Y1", include.name = FALSE, include.wave = TRUE)
}
}
