% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_variables_2w_test.R
\name{get_outcome_better_name}
\alias{get_outcome_better_name}
\title{Get outcome name (better)}
\usage{
get_outcome_better_name(var)
}
\arguments{
\item{var}{a character string (e.g., 'HAPPY_W1')}
}
\value{
a character string
}
\description{
A relatively simple switch function to get a better name for each variable
}
\examples{
{
  get_outcome_better_name("HAPPY_W1")
}
}
