% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_childhood_variables.R
\name{get_variable_codes}
\alias{get_variable_codes}
\title{INCOME',
MARITAL_STATUS_W2_2',
MARITAL_STATUS_W2_4',}
\usage{
get_variable_codes()
}
\description{
INCOME',
MARITAL_STATUS_W2_2',
MARITAL_STATUS_W2_4',
}
