% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_get_variables.R
\name{get_variable_codes}
\alias{get_variable_codes}
\title{Get variable codes other}
\usage{
get_variable_codes(what = "all", appnd = "", ...)
}
\arguments{
\item{what}{to return (options: "all", 'OUTCOME.VEC', 'LIST.COMPOSITES', 'DEMOGRAPHICS.CHILDHOOD.PRED.VEC')}

\item{appnd}{an optional character top append to end of all variable strings, does not apply to composites}

\item{...}{argument not used}
}
\value{
a character string
}
\description{
A simple function to get a list of variable codes used throughout the package
}
\examples{
{
  get_variable_codes()
}
}
