% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_coordinated_tables.R
\name{gfs_generate_main_doc}
\alias{gfs_generate_main_doc}
\title{Construct GFS Main text results}
\usage{
gfs_generate_main_doc(
  df.raw = NULL,
  meta.wopc = NULL,
  meta.wpc = NULL,
  focal.better.name = "Focal Predictor",
  focal.predictor.reference.value = "estimated population mean of focal predictor",
  focal.predictor = NULL,
  p.bonferroni = NULL,
  baseline.pred = NULL,
  outcome.vec = NULL,
  mylabels = NULL,
  res.dir = "results",
  wgt = as.name("WGT"),
  wgt1 = as.name("ANNUAL_WEIGHT_R2"),
  wgt2 = as.name("SAMP.ATTR.WGT"),
  psu = as.name("PSU"),
  strata = as.name("STRATA"),
  ci.bonferroni = FALSE
)
}
\arguments{
\item{df.raw}{a data.frame containing the raw data with values coded as labels}

\item{meta.wopc}{a nested data.frame of results from the meta-analysis function (see gfs_meta_analysis(.)) of the meta-analytic results that did NOT include the principal components in the country-specific regression analyses}

\item{meta.wpc}{a nested data.frame of results from the meta-analysis function (see gfs_meta_analysis(.)) of the meta-analytic results that did INCLUDED the principal components in the country-specific regression analyses}

\item{focal.better.name}{a character that is used as the printed name in tables/captions to denote the focal predictor}

\item{focal.predictor.reference.value}{(character) describing the baseline/reference group for the focal predictor.}

\item{p.bonferroni}{a number (e.g., 0.00081), is internally determined based on number of rows in meta.wopc if not provided}

\item{baseline.pred}{a vector of characters defining which baseline characteristics were used as control variables. This can be used to force the inclusion some variable into the main text summary table.}

\item{outcome.vec}{a character vector of outcomes names (e.g., "HAPPY_Y2") that are to be printed in the main text meta-analytic summary table. Name MUST be included in the meta.wopc (meta.wpc) nested data.frames column (OUTCOME0), otherwise the variable won't be printed.}

\item{mylabels}{an optional character vector that will be printed out in specific rows of tables 2/3 depending on the specification pf outcome.vec}

\item{res.dir}{(character) defaults to "results", and will be created if needed to story results document}

\item{...}{other arguments as needed}
}
\value{
a word document saved to the current 'results/' directory
}
\description{
TO-DO
}
\details{
Generated a word document containing the results for the meta-analytic outcome wide results for the GFS coordinated analyses.
}
\examples{
{
  # TO-DO
}
}
