% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_coordinated_tables.R
\name{gfs_generate_supplemental_docs}
\alias{gfs_generate_supplemental_docs}
\title{Construct GFS Online Supplement results}
\usage{
gfs_generate_supplemental_docs(
  df.raw = NULL,
  focal.predictor = NULL,
  focal.better.name = "Focal Predictor",
  focal.predictor.reference.value = "estimated population mean of focal predictor",
  dir.primary = "results-primary",
  dir.supp = "results-cca",
  dir.attr.models = "results-attr",
  file.primary.wopc = "0_meta_analyzed_results_primary_wopc.rds",
  file.primary.wpc = "0_meta_analyzed_results_primary_wpc.rds",
  file.unstd.wopc = "0_meta_analyzed_results_unstd_wopc.rds",
  file.unstd.wpc = "0_meta_analyzed_results_unstd_wpc.rds",
  file.cca.wopc = "0_meta_analyzed_results_cca_wopc.rds",
  file.cca.wpc = "0_meta_analyzed_results_cca_wpc.rds",
  p.bonferroni = NULL,
  baseline.pred = NULL,
  outcome.vec = NULL,
  mylabels = NULL,
  wgt = WGT0,
  wgt1 = ANNUAL_WEIGHT_R2,
  wgt2 = AVG.SAMP.ATTR.WGT,
  psu = PSU,
  strata = STRATA,
  res.dir = "results",
  included.countries = NULL,
  ci.bonferroni = FALSE,
  num.sequential = FALSE,
  forest.plot.type = "combined",
  what = "all",
  only.figs = FALSE,
  fig.num.start = 0
)
}
\arguments{
\item{df.raw}{a data.frame containing the raw data with values coded as labels}

\item{focal.better.name}{a character that is used as the printed name in tables/captions to denote the focal predictor}

\item{focal.predictor.reference.value}{(character) describing the baseline/reference group for the focal predictor.}

\item{p.bonferroni}{a number (e.g., 0.007), is internally determined based on number of rows in coun.wopc}

\item{baseline.pred}{a vector of characters defining which baseline characteristics were used as control variables. This can be used to force the inclusion some variable into the main text summary table.}

\item{outcome.vec}{a character vector of outcomes names (e.g., "HAPPY_Y2") that are to be printed in the main text meta-analytic summary table. Name MUST be included in the coun.wopc (coun.wpc) nested data.frames column (OUTCOME0), otherwise the variable won't be printed.}

\item{mylabels}{an optional character vector that will be printed out in specific rows of tables 2/3 depending on the specification pf outcome.vec}

\item{res.dir}{(character) defaults to "results", and will be created if needed to story results document}

\item{included.countries}{a character vector of which countries to include in output -- defaults to all.}

\item{what}{(character) options include: "all", "S1", "S2", "S3".}

\item{...}{other arguments as needed}
}
\value{
a word document saved to the current 'results/' directory
}
\description{
Generates three online supplements.

Supplement 1: Supplemental main-text results

(1) Summary statistics of outcomes by wave (raw data + imputed data)
(2) Supplemental meta-analyzed results
   - Compare attrition weights results with multiple-imputation results (one for each model)

Supplement 2: Country-specific results

   - Summary statistics by wave for demographics (similar to main text Table 1)
   - Summary statistics by wave for outcomes (similar to Table S1)
  - Summary of attrition model and distribution of attrition weights
   - Summary statistics of principal components by outcome (# retained, % prop explained, cumsum % prop explained)
   - Outcome-wide results-attrition weights & multiple imputation (similar to main text Table 2-extra-wide p. format)
   - Outcome-wide E-values (similar to main text Table 3)

Supplement 3: Restructured country-specific results + forest plots

(1) Summary statistics of demographics by country (raw data)
(2) Summary statistics of outcomes by country (raw data)
(3) Forest plots of all effects
    - Model 1 (No PCs -- attrition weights)
    - Model 2 (w/ PCs -- attrition weights)
}
\details{
Generated two word documents containing the supplemental results for the outcome wide results for the GFS coordinated analyses.
}
\examples{
{
## Generate online supplement in parts
#
# ## Part 1. Global summary statistics and supplemental meta-analytic results
# # set what = "S1"
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S1"
# )
#
# ## Part 2. Country specific results
# # set what = "S2"
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S2"
# )

## Note. This can also be done is subcomponents if the above crashes
##   Uses the 'included.countries' arguments.

# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S2",
#   included.countries = c('Argentina', 'Australia', 'Brazil', 'China', 'Egypt')
# )
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S2",
#   included.countries = c('Germany', 'Hong Kong', 'India', 'Indonesia', 'Israel', 'Japan', 'Kenya', 'Mexico')
# )
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S2",
#   included.countries = c('Nigeria', 'Philippines', 'Poland', 'South Africa', 'Spain', 'Sweden')
# )
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S2",
#   included.countries = c('Tanzania', 'Turkey', 'United Kingdom', 'United States')
# )

#
# ## Part 3. Summary statistics by country + wave in LARGE tables
# # set what = "S3"
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S3"
# )
#
# ## Part 4. Forest plots -- can be done is pieces as well.
# # set what = "S4"
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S4"
# )

## only a subset of figures at a time
##   Please note: this requires specifying the outcome AND the figure number

## Example 1. Same the above because it uses the entire set of outcomes

# my.vec <- c( "COMPOSITE_FLOURISHING_SECURE_Y2", "COMPOSITE_FLOURISHING_Y2", "COMPOSITE_HAPPI_LIFE_SAT_Y2", "COMPOSITE_HEALTH_Y2", "COMPOSITE_MEANING_PURPOSE_Y2", "COMPOSITE_CHARACTER_Y2", "COMPOSITE_SUBJECTIVE_SOC_CONN_Y2", "COMPOSITE_FINL_MAT_WORRY_Y2", 'HAPPY_Y2',  'LIFE_SAT_Y2', 'WB_TODAY_Y2', 'WB_FIVEYRS_Y2', 'EXPECT_GOOD_Y2', 'FREEDOM_Y2', 'PEACE_Y2', 'LIFE_BALANCE_Y2', 'CAPABLE_Y2', 'WORTHWHILE_Y2', 'LIFE_PURPOSE_Y2', 'MENTAL_HEALTH_Y2', 'THREAT_LIFE_Y2', 'COMPOSITE_DEPRESSION_Y2', 'DEPRESSED_Y2', 'INTEREST_Y2', 'COMPOSITE_ANXIETY_Y2', 'FEEL_ANXIOUS_Y2', 'CONTROL_WORRY_Y2', 'SUFFERING_Y2', 'CONTENT_Y2', 'SAT_RELATNSHP_Y2', 'PEOPLE_HELP_Y2', 'CLOSE_TO_Y2', 'APPROVE_GOVT_Y2', 'SAY_IN_GOVT_Y2', 'BELONGING_Y2', 'SAT_LIVE_Y2', 'TRUST_PEOPLE_Y2', 'MARITAL_STATUS_EVER_MARRIED_Y2', 'MARITAL_STATUS_DIVORCED_Y2', 'NUM_CHILDREN_Y2', 'GROUP_NOT_REL_Y2', 'ATTEND_SVCS_Y2', 'LONELY_Y2', 'DISCRIMINATED_Y2', 'PROMOTE_GOOD_Y2', 'GIVE_UP_Y2', 'HOPE_FUTURE_Y2', 'GRATEFUL_Y2', 'SHOW_LOVE_Y2', 'FORGIVE_Y2', 'DONATED_Y2', 'HELP_STRANGER_Y2', 'VOLUNTEERED_Y2', 'PHYSICAL_HLTH_Y2', 'HEALTH_PROB_Y2', 'BODILY_PAIN_Y2', 'CIGARETTES_BINARY_Y2', 'DRINKS_Y2', 'DAYS_EXERCISE_Y2', 'EXPENSES_Y2', 'WORRY_SAFETY_Y2', 'EDUCATION_3_Y2', 'EMPLOYMENT_Y2', 'INCOME_FEELINGS_Y2', 'OWN_RENT_HOME_Y2', 'INCOME_QUINTILE_Y2', 'CONNECTED_REL_Y2', 'AFTER_DEATH_Y2', 'REL_EXPERIENC_Y2', 'SACRED_TEXTS_Y2', 'PRAY_MEDITATE_Y2', 'BELIEVE_GOD_Y2', 'LIFE_APPROACH_Y2', 'COMFORT_REL_Y2', 'LOVED_BY_GOD_Y2', 'GOD_PUNISH_Y2', 'CRITICAL_Y2', 'TELL_BELIEFS_Y2' )
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S4",
#   outcome.vec = my.vec,
#   fig.num.start = 1
# )

#
# ## Example 2. figures in pieces
# # the first 8 "composite outcomes forest plots were created. The below specification adds the next 12 figures
# my.vec <- c( 'HAPPY_Y2',  'LIFE_SAT_Y2', 'WB_TODAY_Y2', 'WB_FIVEYRS_Y2', 'EXPECT_GOOD_Y2', 'FREEDOM_Y2', 'PEACE_Y2', 'LIFE_BALANCE_Y2', 'CAPABLE_Y2', 'WORTHWHILE_Y2', 'LIFE_PURPOSE_Y2', 'MENTAL_HEALTH_Y2')
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S4",
#   outcome.vec = my.vec,
#   fig.num.start = 9 # 8 were already created, so start with figure number 9
# )
#
# ## create remaining figures
# my.vec <- c('THREAT_LIFE_Y2', 'COMPOSITE_DEPRESSION_Y2', 'DEPRESSED_Y2', 'INTEREST_Y2', 'COMPOSITE_ANXIETY_Y2', 'FEEL_ANXIOUS_Y2', 'CONTROL_WORRY_Y2', 'SUFFERING_Y2', 'CONTENT_Y2', 'SAT_RELATNSHP_Y2', 'PEOPLE_HELP_Y2', 'CLOSE_TO_Y2', 'APPROVE_GOVT_Y2', 'SAY_IN_GOVT_Y2', 'BELONGING_Y2', 'SAT_LIVE_Y2', 'TRUST_PEOPLE_Y2', 'MARITAL_STATUS_EVER_MARRIED_Y2', 'MARITAL_STATUS_DIVORCED_Y2', 'NUM_CHILDREN_Y2', 'GROUP_NOT_REL_Y2', 'ATTEND_SVCS_Y2', 'LONELY_Y2', 'DISCRIMINATED_Y2', 'PROMOTE_GOOD_Y2', 'GIVE_UP_Y2', 'HOPE_FUTURE_Y2', 'GRATEFUL_Y2', 'SHOW_LOVE_Y2', 'FORGIVE_Y2', 'DONATED_Y2', 'HELP_STRANGER_Y2', 'VOLUNTEERED_Y2', 'PHYSICAL_HLTH_Y2', 'HEALTH_PROB_Y2', 'BODILY_PAIN_Y2', 'CIGARETTES_BINARY_Y2', 'DRINKS_Y2', 'DAYS_EXERCISE_Y2', 'EXPENSES_Y2', 'WORRY_SAFETY_Y2', 'EDUCATION_3_Y2', 'EMPLOYMENT_Y2', 'INCOME_FEELINGS_Y2', 'OWN_RENT_HOME_Y2', 'INCOME_QUINTILE_Y2', 'CONNECTED_REL_Y2', 'AFTER_DEATH_Y2', 'REL_EXPERIENC_Y2', 'SACRED_TEXTS_Y2', 'PRAY_MEDITATE_Y2', 'BELIEVE_GOD_Y2', 'LIFE_APPROACH_Y2', 'COMFORT_REL_Y2', 'LOVED_BY_GOD_Y2', 'GOD_PUNISH_Y2', 'CRITICAL_Y2', 'TELL_BELIEFS_Y2' )
# gfs_generate_supplemental_docs(
#   df.raw = df.raw,
#   focal.predictor = FOCAL_PREDICTOR,
#   focal.better.name =  FOCAL_PREDICTOR_BETTER_NAME,
#   focal.predictor.reference.value = FOCAL_PREDICTOR_REFERENCE_VALUE,
#   what = "S4",
#   outcome.vec = my.vec,
#   fig.num.start = 21 # 8 + 12 figures already created, start with 21
# )
#
}
}
