% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_coordinated_tables.R
\name{gfs_generate_supplemental_docs}
\alias{gfs_generate_supplemental_docs}
\title{Construct GFS Online Supplement results}
\usage{
gfs_generate_supplemental_docs(
  df.raw = NULL,
  focal.predictor = NULL,
  focal.better.name = "Focal Predictor",
  focal.predictor.reference.value = "estimated population mean of focal predictor",
  dir.primary = "results-primary",
  dir.supp = "results-cca",
  dir.attr.models = "results-attr",
  file.primary.wopc = "0_meta_analyzed_results_primary_wopc.rds",
  file.primary.wpc = "0_meta_analyzed_results_primary_wpc.rds",
  file.unstd.wopc = "0_meta_analyzed_results_unstd_wopc.rds",
  file.unstd.wpc = "0_meta_analyzed_results_unstd_wpc.rds",
  file.cca.wopc = "0_meta_analyzed_results_cca_wopc.rds",
  file.cca.wpc = "0_meta_analyzed_results_cca_wpc.rds",
  p.bonferroni = NULL,
  baseline.pred = NULL,
  outcome.vec = NULL,
  mylabels = NULL,
  wgt = WGT0,
  wgt1 = ANNUAL_WEIGHT_R2,
  wgt2 = AVG.SAMP.ATTR.WGT,
  psu = PSU,
  strata = STRATA,
  res.dir = "results",
  included.countries = NULL,
  ci.bonferroni = FALSE,
  num.sequential = FALSE,
  what = "all"
)
}
\arguments{
\item{df.raw}{a data.frame containing the raw data with values coded as labels}

\item{focal.better.name}{a character that is used as the printed name in tables/captions to denote the focal predictor}

\item{focal.predictor.reference.value}{(character) describing the baseline/reference group for the focal predictor.}

\item{p.bonferroni}{a number (e.g., 0.007), is internally determined based on number of rows in coun.wopc}

\item{baseline.pred}{a vector of characters defining which baseline characteristics were used as control variables. This can be used to force the inclusion some variable into the main text summary table.}

\item{outcome.vec}{a character vector of outcomes names (e.g., "HAPPY_Y2") that are to be printed in the main text meta-analytic summary table. Name MUST be included in the coun.wopc (coun.wpc) nested data.frames column (OUTCOME0), otherwise the variable won't be printed.}

\item{mylabels}{an optional character vector that will be printed out in specific rows of tables 2/3 depending on the specification pf outcome.vec}

\item{res.dir}{(character) defaults to "results", and will be created if needed to story results document}

\item{included.countries}{a character vector of which countries to include in output -- defaults to all.}

\item{what}{(character) options include: "all", "S1", "S2", "S3".}

\item{...}{other arguments as needed}
}
\value{
a word document saved to the current 'results/' directory
}
\description{
Generates three online supplements.

Supplement 1: Supplemental main-text results

(1) Summary statistics of outcomes by wave (raw data + imputed data)
(2) Supplemental meta-analyzed results
   - Compare attrition weights results with multiple-imputation results (one for each model)

Supplement 2: Country-specific results

   - Summary statistics by wave for demographics (similar to main text Table 1)
   - Summary statistics by wave for outcomes (similar to Table S1)
  - Summary of attrition model and distribution of attrition weights
   - Summary statistics of principal components by outcome (# retained, % prop explained, cumsum % prop explained)
   - Outcome-wide results-attrition weights & multiple imputation (similar to main text Table 2-extra-wide p. format)
   - Outcome-wide E-values (similar to main text Table 3)

Supplement 3: Restructured country-specific results + forest plots

(1) Summary statistics of demographics by country (raw data)
(2) Summary statistics of outcomes by country (raw data)
(3) Forest plots of all effects
    - Model 1 (No PCs -- attrition weights)
    - Model 2 (w/ PCs -- attrition weights)
}
\details{
Generated two word documents containing the supplemental results for the outcome wide results for the GFS coordinated analyses.
}
\examples{
{
  # TO-DO
}
}
