% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfs_meta_analysis.R
\name{gfs_meta_analysis}
\alias{gfs_meta_analysis}
\title{GFS Wrapper for Meta-Analysis}
\usage{
gfs_meta_analysis(
  meta.input,
  estimator = "PM",
  interval.method = "empirical",
  better.name = NULL,
  p.subtitle = NULL,
  ...
)
}
\arguments{
\item{meta.input}{a nested data.frame where `data` is a column}

\item{...}{additional arguments}
}
\value{
a list of meta-analysis results from the metafor package
}
\description{
Conducted meta-analysis (random effects or population weighted) for a set of effect of interest.
}
\examples{
# TO-DO
}
