% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoding_imputed_data.R
\name{recode_imputed_data}
\alias{recode_imputed_data}
\title{Reformat Imputed Data}
\usage{
recode_imputed_data(
  df.imp = NULL,
  list.default = NULL,
  list.composites = NULL,
  list.manual = NULL,
  wave = 2,
  method.income = "quintiles.top.fixed",
  ...
)
}
\arguments{
\item{df.imp}{a nested data.frame with results from run_impute_data(.).}

\item{list.default}{(required) a names list with DEMOGRAPHICS.CHILDHOOD.PRED.VEC, OUTCOME.VE,
FOCAL_PREDICTOR, USE_DEFAULT, FORCE_BINARY, VALUES_DEFINING_UPPER_CATEGORY, VALUES_DEFINING_LOWER_CATEGORY, FORCE_CONTINUOUS}

\item{list.composites}{(optional) a named list with elements LIST.OUTCOME.COMPOSITES, LIST.COMPOSITE.COMBINE.METHOD and COMPOSITE.VEC.}

\item{list.manual}{(optional) a list like list.default (to-do)}

\item{wave}{(default is wave 2) but can be coerced to use wave 1 data (use wave = 1 to appropriately utilize recoding of wave 1 data)}

\item{method.income}{(optional) method for how income, based on country specific labels, is recoded. Options include 'quintiles.num.fixed', 'quintiles.num.random', 'quintiles.top.fixed', ''quintiles.top.random', 'numeric'.}

\item{...}{other arguments passed to mice}
}
\value{
a long data.frame for use in analyses
}
\description{
By default, the recoding creates a set of new variables for each demographic and childhood predictor variable with the prefix "COV_".

method.income has several options.
* 'quintiles.num.fixed', 'quintiles.num.random' use survey::svyquantile() to identify the country specific categories that separate responses into quintiles. This is only approximate however because the categories are coarse and the top 20% isn't guarenteed to be exact. The used of the 'fixed' version uses these coarse not exactly 20% quantiles. The 'random' version will randomly sample exactly 20% from a top down approach to obtain quintiles.
* 'quintiles.top.fixed', ''quintiles.top.random' use the
* 'numeric

More to come...
}
\details{
Reformated the imputed (nested) data for use in coordinated regression analyses.
}
\examples{
{
  # none
}
}
