% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variable_value_switch_funtions.R
\name{recode_labels}
\alias{recode_labels}
\title{Recode Variable Labels}
\usage{
recode_labels(x, var, include.value = TRUE, add.whitespace = FALSE)
}
\arguments{
\item{x}{a vector}

\item{var}{a character string (e.g., 'HAPPY_W1')}

\item{include.value}{a logical (default = TRUE) of whether to include the numeric value in the returned label. Not applicable for numeric variables.}

\item{add.whitespace}{a logical (efault = FALSE) of whether to add whitespace to beginning of label (makes summary tables easier to read).}
}
\value{
a vector
}
\description{
A relatively simple switch function to transform variables to from the numeric values
to the labels associated with those levels.
}
\examples{
{
  # TO-DO
}
}
