% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_wgts.R
\name{run_attrition_model_by_country}
\alias{run_attrition_model_by_country}
\title{Wrapper Function Attrition Weights}
\usage{
run_attrition_model_by_country(
  data.dir,
  wgt = "ANNUAL_WEIGHT_R2",
  attr.wgt.name = NULL,
  pooled.wgt = NULL,
  composite.wgt.name = NULL,
  ...
)
}
\arguments{
\item{data.dir}{character defining where the recoded imputed data files are stored}

\item{...}{additional arguments passed to ....}
}
\value{
a dataset with attrition weights appended
}
\description{
TO-DO
}
\details{
Wrapper function to use the data.set specific attrition weights function.
}
\examples{
{
  # TO-DO
}
}
