% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_impute_data.R
\name{run_impute_data}
\alias{run_impute_data}
\title{Impute Missing GFS Data}
\usage{
run_impute_data(data, data.dir, Nimp = 20, Miter = 5, ...)
}
\arguments{
\item{data}{data.frame with raw data with missing coded as "(missing)"}

\item{data.dir}{directory where to save the imputed data}

\item{Nimp}{number of imputed datasets to generated}

\item{Miter}{number of iteration per imputed dataset to use during imputatiuon (uses mice default)}

\item{...}{other arguments passed to mice}
}
\value{
a nested (by country) data.frame of mice objects with completed imputed datasets ready
for extraction and recoding.
}
\description{
Additional details forthcoming....

The imputation automatically excludes variables that are 100% in some countries (e.g., APPROVE_GOVT in Egypt) and variables that are conditional on previous questions (e.g, REL3-9 and TEACHINGS1-15).

The imputation method for all variables is predictive mean matching (pmm).
}
\details{
A wrapper function to conducting imputation by country.
}
\examples{
{
  # TO-DO
}
}
